+++ NEWS.md
@@ -1,6 +1,7 @@
 # ggblanket 12.3.0
 
 * Breaking: renamed `aes_colour_contrast` to `aes_contrast`.
+* Breaking: removed `linewidthness` helper.
 * Support default titles, subtitles and captions in `set_blanket`.
 * Changed default `*_breaks_n` for `x` or `y` to be either 6 or 4 (if faceted). 
 Note, symmetric uses `only.loose = TRUE` whereas `FALSE` for non-symmetric.